toggle visible/hidden with current parcel scene state
