+++ rust/decentraland-godot-lib/src/scene_runner/components/ui/mod.rs
@@ -1,5 +1,5 @@
 pub mod scene_ui;
-pub mod taffy;
+pub mod style;
 pub mod ui_background;
 pub mod ui_text;
 pub mod ui_transform;
