+++ rust/decentraland-godot-lib/src/scene_runner/scene.rs
@@ -3,7 +3,10 @@ use std::{
     time::Instant,
 };
 
-use godot::prelude::{Dictionary, Gd};
+use godot::{
+    engine::Control,
+    prelude::{Dictionary, Gd},
+};
 
 use crate::{
     common::rpc::RpcCalls,
@@ -200,8 +203,9 @@ impl Scene {
         dcl_scene: DclScene,
         content_mapping: Dictionary,
         scene_type: SceneType,
+        parent_ui_node: Gd<Control>,
     ) -> Self {
-        let godot_dcl_scene = GodotDclScene::new(&scene_definition, &scene_id);
+        let godot_dcl_scene = GodotDclScene::new(&scene_definition, &scene_id, parent_ui_node);
 
         Self {
             scene_id,
@@ -253,7 +257,8 @@ impl Scene {
         let scene_id = Scene::new_id();
         let dcl_scene = DclScene::spawn_new_test_scene(scene_id);
         let content_mapping = Dictionary::default();
-        let godot_dcl_scene = GodotDclScene::new(&scene_definition, &scene_id);
+        let godot_dcl_scene =
+            GodotDclScene::new(&scene_definition, &scene_id, Control::new_alloc());
 
         Self {
             scene_id,
